updated tests and training printouts, moved downsample()
