+++ xrsdkit/models/__init__.py
@@ -4,13 +4,14 @@ from collections import OrderedDict
 
 import yaml
 import pandas as pd
+import numpy as np
 from citrination_client import CitrinationClient
+from sklearn import preprocessing
 
 from .. import definitions as xrsdefs 
 from .regressor import Regressor
 from .classifier import Classifier
-from ..tools import primitives
-from ..tools.citrination_tools import downsample, get_data_from_Citrination
+from ..tools import primitives, profiler, citrination_tools 
 
 file_path = os.path.abspath(__file__)
 src_dir = os.path.dirname(os.path.dirname(file_path))
@@ -167,7 +168,7 @@ def downsample_and_train(
         if True, the downsampling statistics and models will be
         saved in modeling_data/testing_data dir
     """
-    df = get_data_from_Citrination(citrination_client,source_dataset_ids)
+    df = citrination_tools.get_data_from_Citrination(citrination_client,source_dataset_ids)
     df_sample = downsample_by_group(df)
     train_from_dataframe(df_sample,train_hyperparameters,save_models,test)
 
@@ -197,10 +198,85 @@ def downsample_by_group(df):
     #for exp_id in all_exp:
     # downsample each group independently
     for group_labels,grp in all_groups.groups.items():
+        group_df = df.iloc[grp].copy()
+        print('Downsampling data for group: {}'.format(group_labels))
         #lbl_df = _filter_by_labels(data,lbls)
         dsamp = downsample(df.iloc[grp].copy(), 1.0)
+        print('Finished downsampling: kept {}/{}'.format(len(dsamp),len(group_df)))
         data_sample = data_sample.append(dsamp)
     return data_sample
+
+def downsample(df, min_distance):
+    """Downsample records from one DataFrame.
+
+    Transforms the DataFrame feature arrays 
+    (scaling by the columns in profiler.profile_keys),
+    before collecting at least 10 samples.
+    If the size of `df` is <= 10, it is returned directly.
+    If it is larger than 10, the first point is chosen
+    based on greatest nearest-neighbor distance.
+    Subsequent points are chosen  
+    in order of decreasing nearest-neighbor distance
+    to the already-sampled points. 
+
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        dataframe containing xrsd samples 
+    min_distance : float
+        the minimum allowed nearest-neighbor distance 
+        for continuing to downsample after 10 or more samples
+        have been selected 
+
+    Returns
+    -------
+    sample : pandas.DataFrame
+        dataframe containing subset of rows
+        that was chosen using distance between the samples
+    """
+    df_size = len(df)
+    sample = pd.DataFrame(columns=df.columns)
+    if df_size <= 10:
+        sample = sample.append(df)
+    else:
+        features = profiler.profile_keys
+        scaler = preprocessing.StandardScaler()
+        scaler.fit(df[features])
+
+        features_matr = scaler.transform(df[features]) # features_matr is a np arraly
+        # define the distance between two samples in feature space
+
+        dist_func = lambda i,j: np.sum(
+            np.abs(features_matr[i]
+            - features_matr[j]))
+        dist_matrix = np.array([[dist_func(i,j) for i in range(df_size)] for j in range(df_size)])
+        # get the most isolated sample first:
+        # this should be the sample with the greatest 
+        # nearest-neighbor distance 
+        nn_distance_array = np.array([min(dist_matrix[i,:]) for i in range(df_size)])
+        best_idx = np.argmax(nn_distance_array)
+        sample = sample.append(df.iloc[best_idx])
+        sampled_idxs = [best_idx] 
+        continue_downsampling = True
+        while(continue_downsampling):
+            # find the next best index to sample:
+            # the sample with the greatest minimum distance
+            # between itself and the downsampled samples
+            sample_size = len(sample)
+            sample_dist_matrix = np.array([dist_matrix[i,:] for i in sampled_idxs])
+            nn_distance_array = np.array([min(sample_dist_matrix[:,j]) for j in range(df_size)])
+            best_idx = np.argmax(nn_distance_array)
+            best_nn_distance = nn_distance_array[best_idx]
+            # if we have at least 10 samples,
+            # and all remaining samples are close to the current sample,
+            # down-sampling can stop here.
+            if sample_size >= 10 and best_nn_distance < min_distance: 
+                continue_downsampling = False
+            else:
+                sampled_idxs.append(best_idx)
+                sample = sample.append(df.iloc[best_idx])
+    return sample
+
 
 def train_from_dataframe(data,train_hyperparameters=False,save_models=False,test=False):
     # regression models:
@@ -252,6 +328,8 @@ def train_regression_models(data, hyper_parameters_search=False):
                     except:
                         pass
                     reg_model.train(sys_cls_data, hyper_parameters_search)
+                    if not reg_model.trained: 
+                        print('    training failed: insufficient data or zero variance')
                     pop_models[k] = reg_model 
                 elif k in xrsdefs.setting_selections['lattice']:
                     print('    structure: {}'.format(k))
@@ -269,6 +347,8 @@ def train_regression_models(data, hyper_parameters_search=False):
                         except:
                             pass
                         reg_model.train(sub_cls_data, hyper_parameters_search)
+                        if not reg_model.trained: 
+                            print('        training failed: insufficient data or zero variance')
                         pop_models[k][param_nm] = reg_model 
                 elif k in xrsdefs.setting_selections['interaction']:
                     print('    interaction: {}'.format(k))
@@ -286,6 +366,8 @@ def train_regression_models(data, hyper_parameters_search=False):
                         except:
                             pass
                         reg_model.train(sub_cls_data, hyper_parameters_search)
+                        if not reg_model.trained: 
+                            print('        training failed: insufficient data or zero variance')
                         pop_models[k][param_nm] = reg_model 
                 else:
                     # k is a basis classification
@@ -307,6 +389,8 @@ def train_regression_models(data, hyper_parameters_search=False):
                             except:
                                 pass
                             reg_model.train(bas_cls_data, hyper_parameters_search)
+                            if not reg_model.trained: 
+                                print('            training failed: insufficient data or zero variance')
                             specie_models[param_nm] = reg_model 
     return reg_models
 
@@ -410,29 +494,38 @@ def train_classification_models(data, hyper_parameters_search=False):
         trained on the given dataset `data`.
     """
     cls_models = trainable_classification_models(data)
+    sys_cls_model = cls_models.pop('system_classification')
+    if sys_cls_model:
+        print(os.linesep+'Training main system classifier')
+        model = Classifier('system_classification',None)
+
+        if 'system_classification'in classification_models.keys(): # we have a trained model
+            old_pars = classification_models['system_classification'].model.get_params()
+            model.model.set_params(alpha=old_pars['alpha'], l1_ratio=old_pars['l1_ratio'])
+
+        model.train(data, hyper_parameters_search=hyper_parameters_search)
+        if not model.trained: 
+            print('insufficient or uniform training data: using default value')
+        cls_models['system_classification'] = model
     for sys_cls_lbl, sys_models in cls_models.items():
-        if sys_cls_lbl == 'system_classification':
-            model = Classifier('system_classification',None)
-
-            if 'system_classification'in classification_models.keys(): # we have a trained model
-                old_pars = classification_models['system_classification'].model.get_params()
-                model.model.set_params(alpha=old_pars['alpha'], l1_ratio=old_pars['l1_ratio'])
-
-            model.train(data, hyper_parameters_search=hyper_parameters_search)
-            cls_models['system_classification'] = model
-        else:
-            for pop_id,pop_models in sys_models.items():
-                for cls_label in pop_models.keys():
-                    model_label = pop_id+'_'+cls_label
-                    data_subset = data[data['system_classification']==sys_cls_lbl]
-                    m = Classifier(model_label,None)
-                    try: # check if we alredy have a trained model for this label
-                        old_pars = classification_models[sys_cls_lbl][pop_id][cls_label].model.get_params()
-                        m.model.set_params(alpha=old_pars['alpha'], l1_ratio=old_pars['l1_ratio'])
-                    except:
-                        pass
-                    m.train(data_subset, hyper_parameters_search=hyper_parameters_search)
-                    pop_models[cls_label] = m 
+        print('Training classifiers for system: ')
+        print(sys_cls_lbl)
+        for pop_id,pop_models in sys_models.items():
+            print('population id: {}'.format(pop_id))
+            for cls_label in pop_models.keys():
+                model_label = pop_id+'_'+cls_label
+                print('    classifier: {}'.format(model_label))
+                data_subset = data[data['system_classification']==sys_cls_lbl]
+                m = Classifier(model_label,None)
+                try: # check if we alredy have a trained model for this label
+                    old_pars = classification_models[sys_cls_lbl][pop_id][cls_label].model.get_params()
+                    m.model.set_params(alpha=old_pars['alpha'], l1_ratio=old_pars['l1_ratio'])
+                except:
+                    pass
+                m.train(data_subset, hyper_parameters_search=hyper_parameters_search)
+                if not m.trained: 
+                    print('    insufficient or uniform training data: using default value')
+                pop_models[cls_label] = m 
     return cls_models
 
 def trainable_classification_models(data):
